Add files via upload
Wordle-PyGame-main  directory
This directory is a sample code for pygame that provides a python file which provides a wordle pygame interface.
This code was pulled from Github repository: https://github.com/baraltech/Wordle-PyGame and cited in the code detailing the original author's contributions.  Each line of code has been commented such that as the code is modified the original author's contributions and our modifications can be differentiated.

The next step is to modify this code to provide a play step such that the agent process can directly interact with the pygame interface directly in the while true loop that is the game loop.  The difficulty in the pygame interface is that it is a window interface that accepts keyboard keystrokes as input into the game.  I have experimented with python interactions that should provide keyboard emulation but the attempts so far have failed.  

therefore the approach to embedding the guessing agent into the while true loop should provide dirct interactions with the pygame interface without the need to resort to the keyboard interface.