Add files via upload
I wrote a feedback function that improves on the feedback capabilities from the main script
This explores the feedback function and shows how to use it in multiple configurations.
The main purpose the feedback provides is information for the guessing agent.
there are multiple products provided in this script.
one major capability is the ability to evaluate the vocabulary list in a hypothetical context
using the feedback in one mode can evaluate a guess against a proposed secret word and calculate: given a hypothetical secret word, evaluate each vocabulary word to determine what the length of the resulting vocabulary would be if each word were the guess.  
running the script in a np.vectorize context provides a numpy array of vocabulary lengths for all words in the vocabulary.
simply iterate over the entire vocabulary and t=you get a complete evaluation of the entire data set.  

Another mode of the feedback script provides a detailed feedback set of the guess with results provided as a named tuple.